+++ pentobi/qml/Main.qml
@@ -33,8 +33,6 @@ ApplicationWindow {
 
     property bool isWindows: Qt.platform.os === "windows"
 
-    property bool useNativeFileDialog: isWindows
-
     property string themeName: isAndroid ? "dark" : "system"
 
     property QtObject theme: Logic.createTheme(themeName)
@@ -175,25 +173,13 @@ ApplicationWindow {
     DialogLoader { id: gameInfoDialog; url: "GameInfoDialog.qml" }
     DialogLoader { id: initialRatingDialog; url: "InitialRatingDialog.qml" }
     DialogLoader { id: newFolderDialog; url: "NewFolderDialog.qml" }
-    DialogLoader {
-        id: openDialog;
-        url: useNativeFileDialog ? "OpenDialogNative.qml" : "OpenDialog.qml"
-    }
+    DialogLoader { id: openDialog; url: "OpenDialog.qml" }
     DialogLoader { id: exportImageDialog; url: "ExportImageDialog.qml" }
-    DialogLoader {
-        id: imageSaveDialog;
-        url: useNativeFileDialog ? "ImageSaveDialogNative.qml" : "ImageSaveDialog.qml"
-    }
-    DialogLoader {
-        id: asciiArtSaveDialog;
-        url: useNativeFileDialog ? "AsciiArtSaveDialogNative.qml" : "AsciiArtSaveDialog.qml"
-    }
+    DialogLoader { id: imageSaveDialog; url: "ImageSaveDialog.qml" }
+    DialogLoader { id: asciiArtSaveDialog; url: "AsciiArtSaveDialog.qml" }
     DialogLoader { id: gotoMoveDialog; url: "GotoMoveDialog.qml" }
     DialogLoader { id: ratingDialog; url: "RatingDialog.qml" }
-    DialogLoader {
-        id: saveDialog;
-        url: useNativeFileDialog ? "SaveDialogNative.qml" : "SaveDialog.qml"
-    }
+    DialogLoader { id: saveDialog; url: "SaveDialog.qml" }
     DialogLoader { id: infoMessage; url: "MessageDialog.qml" }
     DialogLoader { id: questionMessage; url: "QuestionDialog.qml" }
     DialogLoader { id: analyzeDialog; url: "AnalyzeDialog.qml" }