+++ pentobi/qml/MenuExport.qml
@@ -20,10 +20,6 @@ PentobiMenu {
                 androidUtils.openTextSaveDialog()
             else {
                 var dialog = asciiArtSaveDialog.get()
-                if (! rootWindow.useNativeFileDialog) {
-                    dialog.name = gameModel.suggestFileName(folder, "txt")
-                    dialog.selectNameFilter(0)
-                }
                 dialog.open()
             }
     }