+++ NEWS.md
@@ -4,6 +4,9 @@ Pentobi Release Notes
 Version 23.0-dev (master)
 -------------------------
 
+* Use native file dialogs on Linux to make saving files work in flatpak
+  sandboxes. This adds a run-time dependency on package
+  qml6-module-qtquick-dialogs.
 * Removed option to disable animations.
 
 Version 22.2-dev (v22-fixes)